Rename and reconfigure manifests
