+++ test/integration/multicluster/statefulset-target/testdata/nginx.yaml
@@ -1,8 +1,11 @@
+
 apiVersion: v1
 kind: Service
 metadata:
-  name: nginx-svc
-  namespace: default
+  name: nginx-statefulset-svc
+  namespace: multicluster-statefulset
+  labels:
+    mirror.linkerd.io/exported: "true"
 spec:
   clusterIP: None
   clusterIPs:
@@ -22,9 +25,10 @@ spec:
 apiVersion: apps/v1
 kind: StatefulSet
 metadata:
-  name: nginx-set
+  name: nginx-statefulset
+  namespace: multicluster-statefulset
 spec:
-  serviceName: nginx-svc
+  serviceName: nginx-statefulset-svc
   selector:
     matchLabels:
       app: nginx-set
